Rename node -> polkadot_node; codec -> polkadot_codec
