+++ core/storage/merkle/polkadot_trie_db/CMakeLists.txt
@@ -2,8 +2,8 @@
 # SPDX-License-Identifier: Apache-2.0
 
 add_library(polkadot_trie_db_codec
-    codec.cpp
-    node.cpp
+    polkadot_codec.cpp
+    polkadot_node.cpp
     )
 target_link_libraries(polkadot_trie_db_codec
     buffer
