+++ core/storage/merkle/polkadot_trie_db/polkadot_node.cpp
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#include "storage/merkle/polkadot_trie_db/node.hpp"
+#include "storage/merkle/polkadot_trie_db/polkadot_node.hpp"
 
 namespace kagome::storage::merkle {
